requirements.txt
pygame>=2.0.0
numpy>=1.21.0
PyOpenGL>=3.1.5
pyrr>=0.10.3
tqdm